pyproject.toml
[project]
name = "edgeff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tensorflow",
    "matplotlib",
    "six",
    "torch",
    "torchaudio",
    "torchvision",
    "python-dotenv",
    "wandb",
    "weave",
    "pandas>=3.0.0",
]
